refactor: Move scoreboard event to the right place
