+++ common/src/main/java/com/wynntils/handlers/scoreboard/events/ScoreboardSegmentAdditionEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.event;
+package com.wynntils.handlers.scoreboard.events;
 
 import com.wynntils.core.events.EventThread;
 import com.wynntils.handlers.scoreboard.Segment;
